ajout de map par fichier + supression de map

+++ js/application/index.js
@@ -13,12 +13,21 @@ function volume()
 }
 
 function newGame() {
-    window.localStorage.setItem('loadSavedGame', 'false');
-    window.location.href='html/game.html';
+    if(window.localStorage.getItem('backup') !== null && JSON.parse(window.localStorage.getItem('backup')) != null) {
+        let bool =confirm("Si vous démarrez une nouvelle partie, vous perdrez votre sauvegarde précédente.");
+        if (bool == true) { 
+            window.localStorage.setItem('loadSavedGame', 'false');
+            window.location.href='html/game.html';
+        }
+    }
+    else {
+        window.localStorage.setItem('loadSavedGame', 'false');
+        window.location.href='html/game.html';
+    }
 }
 
 function loadSavedGame() {
-    if(window.localStorage.getItem('map') !== null && JSON.parse(window.localStorage.getItem('map')) != null) 
+    if(window.localStorage.getItem('backup') !== null && JSON.parse(window.localStorage.getItem('backup')) != null) 
     {
         window.localStorage.setItem('loadSavedGame', 'true');
         window.location.href='html/game.html';
